Refactor Video
Add Controller Mask
