--- components/Video/Controller.tsx
@@ -1,12 +1,0 @@
-interface Props {
-  onSwipeUp: () => void;
-  onSwipeDown: () => void;
-  onTap: () => void;
-  onDoubleTap: () => void;
-}
-
-const Controller: React.FC<Props> = ({ onSwipeUp, onSwipeDown, onTap, onDoubleTap }) => {
-  return <div></div>;
-};
-
-export default Controller;
